requirements.txt
docx2txt==0.8
faiss-cpu==1.7.1
langchain==0.1.0
openai==0.27.0
Pillow==9.2.0
pytesseract==0.3.8
python-dotenv==0.20.0
requests==2.28.1
PyMuPDF==1.19.6
streamlit==1.10.0
tiktoken==0.2.0
